+++ app/src/main/java/org/breezyweather/settings/compose/WeatherSourcesSettingsScreen.kt
@@ -157,6 +157,7 @@ fun WeatherSourcesSettingsScreen(
                                         titleId = id,
                                         summary = preference.summary,
                                         content = preference.content,
+                                        placeholder = preference.placeholder,
                                         regex = preference.regex,
                                         regexError = preference.regexError,
                                         onValueChanged = preference.onValueChanged